requirements.txt
email-validator==2.3.0
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gigachat==0.2.0
python-dotenv==1.2.2
requests==2.33.1
SQLAlchemy==2.0.48
